Add migrations, make changes to APIs for jwt required

+++ resources/item.py
@@ -1,7 +1,7 @@
 from models import ItemModel
 from db import db
 from sqlalchemy.exc import SQLAlchemyError
-from flask_jwt_extended import jwt_required
+from flask_jwt_extended import jwt_required, get_jwt
 from flask.views import MethodView
 from flask_smorest import abort, Blueprint
 from schemas import ItemSchema, ItemUpdateSchema
@@ -29,9 +29,14 @@ class Item(MethodView):
         
         db.session.add(item)
         db.session.commit()
+
     
     @jwt_required()
     def delete(self, item_id):
+        jwt = get_jwt()
+        print(jwt)
+        if not jwt.get("is_admin"):
+            abort(401, message="Admin access required.")
         item = ItemModel.query.get_or_404(item_id)
         db.session.delete(item)
         db.session.commit()
@@ -44,7 +49,7 @@ class ItemList(MethodView):
     def get(self):
         return ItemModel.query.all()
     
-    @jwt_required()
+    @jwt_required(fresh=True)
     @blp.arguments(ItemSchema)
     @blp.response(201, ItemSchema)
     def post(self, item_data):
